Add support for including categories within `/location_free_spaces` API endpoint The PMU dashboard of free spaces per location should be broken down by category. Reference data for each category and associated mapping of prisons to categories has been completed previously. This PR leverages that data and adds support for `includes=category` parameter.
Note that the `Locations::Finder` service has been updated to follow the new pattern from `Moves::Finder` of eager fetching only the required relations to match the requested list of includes - this both avoids an N+1 query when categories are included, and needlessly fetching all categories when they are not included. Also added support for sorting by category so that the data can be returned to the front end in a sensible order.