Removes now-unused dynamic presenter
jira link: https://dsdmoj.atlassian.net/browse/P4-2422

We ended up leaning on our json:api serializer gem to render the
relationships. This enables us to automatically support includes on each
of the anonymous serializers and is significantly cleaner.

This does mean that the method removed in this commit is no longer used.