P4-2520 Add editable field on PER
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2520

Surface value to indicate if a PER's responses are editable or not.
A PER is editable if it has not been confirmed, and the move status
is either booked or requested.

For older PERs that do not have moves associated, support
editable on confirmed/unconfirmed only.