P4-2415 Add Youth Risk Assessments Serializers
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2451

Both PER and youth assessments share serializer behaviour, bar the
type returned. To avoid duplication, add a new serializer for framework
assessment and framework assessments which holds most of the logic. Inherit
from each of those serializers but set the type of assessment on the PER
and Youth risk assessment.

The prefill source serializer type cannot be passed in and does not accept
a block, so we need to follow the same approach around inheriting from a
shared prefill source for both the PER and youth assessment.