P4-2378 Add `requires_youth_risk_assessment` attribute to profiles
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2378

If a person is moving from a YOI, the client is required to ask the user
if the person needs a youth risk assessment if their age is between 18 and 19.
The answer can be stored on the profile as a new attribute which is a boolean
representing if the user selects true or false. The attribute remains nil if the
question is not answered.

The logic around when to ask this question will be with the client, which
will be looking at the move from location, and at the age of the person,
being between 18 - 19.