P4-2589 Add new column value_type to responses
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2589

Add new value_type on responses to avoid reaching into questions
everytime in the response serializer. This is a step closer to avoid
eager loading a set of resources everytime we call the PER serializer.

Also add a task to backfill all value types on responses, which queries and
updates values in batches in 1000 responses.