Determine unlock and discharges from Nomis for new population endpoint Previously these values were copied from the previous days population record, but the values can be determined via Nomis instead which should be up to date. Unfortunately given the limitations of the NOMIS API this means making three separate API calls synchronously and then combining the output of these together to get the correct values. NOMIS should be fast enough in production to do this while the front end is waiting for the response - during dev testing the endpoints are reasonably fast, and production is thought to be faster.
I did explore options for making these 3 calls in parallel, but that's actually not a trivial task and is best achieved with one of a few gems designed for asynchronous HTTP calls. On balance it didn't seem worthwhile to add another dependency; if the observed performance is unsatisfactory then this can be revisited. This is a nice reference I found: https://pawelurbanek.com/ruby-concurrent-requests