Fix move filters
If a move filter is present but empty in a query parameter, it is
interpretted as an empty string. When splitting an empty string
the result is an empty array. This means that we query an empty
array instead of nil for filters which does not return the intended
results. Instead we should return nil when the string is empty
to query any filter and expect values that are nil to be returned.