Add webhook and email notifications for PER amendments This requires a new Gov notify template as the email content is different than for moves, and updates to the webhook serializer as we want to send notifications (including a link) related directly to the PER. This presents some interesting challenges as there is a lot of shared behaviour for both notifications and we also need to be careful not to break and notification jobs that may be currently queued in sidekiq, so cannot change any method signatures or underlying behaviour.
The approach I've taken is to add a new mailer class for the PER, but refactored the existing MoveMailer to extract common personalisation attributes to a new BaseMailer. The new PER mailer then needs to be called rather than the MoveMailer from the Notification class; that already has support for polymorphic attachment to any model, but was hardcoded to always use the MoveMailer. That has been refactored to support invoking the appropriate mailer based on the notifiable topic (either a Move or PersonEscortRecord).

Notification jobs for all record types are ultimately delegated to PrepareMoveNotificationsJob - this too makes the assumption that the underlying topic is always a move. The new PreparePersonEscortRecordNotificationsJob descends from that class and overrides two key methods (without changing the method signature for backwards sidekiq compatibility) to resolve the correct PER and the move associated with the PER.

The Notifier service has been updated to use the new PER job only for the case of amendments to a completed PER. For other assessment changes (PER or Youth Risk) the original PrepareAssessmentNotificationsJob is retained (which ultimately delegates to a move notification). This is slightly confusing and perhaps something that will be changed over time, but I didn't want to introduce a breaking change to the existing PER webhooks and potentially break any queued sidekiq jobs.

Finally the FrameworkResponsesController is updated to invoke notifications for PER amendments (i.e. changes to the responses for a completed PER) - we can detect these when the PER has a populated `amended_at` attribute. Youth Risk Assessments do not have this attribute but share the same controller for updating responses.

In a follow up PR the legacy ENV var for GOVUK_NOTIFY_TEMPLATE_ID will be removed (we have two templates now so it makes sense to rename this). The deploy repo already populates correct values for the two new ENV vars - see https://github.com/ministryofjustice/hmpps-book-secure-move-api-deploy/pull/127