Add a test for when a profile has no category
Category is an optional association on a profile, but we don't have a
test for what happens when a move is created with a profile which
doesn't have one. This test shows that currently the move is created
(with a 201 status code being returned).

This test is effectively the same as
https://github.com/ministryofjustice/hmpps-book-secure-move-api/blob/2d712137bb42b96bf83813326c33c97b20ea47a0/spec/requests/api/moves_controller_create_v2_spec.rb#L65-L67
but I think that's fine as it's more explicit and the behaviour is well
defined. If we decided to change the behaviour in the future such that
a category was always required, we may decide to change the first
`let(:profile`) in this test file to be one with a supported category
(and then remove the test lower down in the file which tests that
explicitly).