Add shared example for capturing Sentry message
This commit adds a new RSpec shared example for ensuring that a
`subject` is capturing a message in Sentry. I've also refactored the few
tests which did this to use the new example.

This reduces the amount of set up we need to do in the tests which check
for this, reducing the lines of code and focusing on only the important
parts of the test itself.