Extract developer message from Prison API errors
This extracts the developer message in an error response from the Prison
API into its own piece of additional data for Sentry which should make
it easier to understand the information provided in the error.

This is a nice to have, and to ensure it doesn't ever prevent errors
ending up in Sentry, if the body cannot be parsed or the field can't be
pulled out, the data will be left as `nil` and the raw body continues to
be included in the additional data.