Add assessment dates to framework NOMIS Mappings
Assessments have two date fields that don't have a suitable column in
the database: `approval_date` and `next_review_date`. In theory these
could be mapped to `start_date` and `end_date` but we've already got a
pattern of having date fields named approriate for the type so I thought
it was best to continue that.