Add NotificationFailedResponseError
This is a new error which is raised when a notification webhook returns
with a non-success response. I've introduced this so we can better
organise the errors in Sentry as there will be a clear exception. It
also allows us to record these as warnings in Sentry, rather than
errors.