Remove Rack Timeout
This was added to monitor timeouts six months ago, but I would argue
there are better ways of monitoring this.

Firstly, Sentry has a relatively new "performance" section which,
among other things, allows users to see the request duration and
trace where slow parts of the request are. I believe that using Rack
Timeout is conflicting with the performance tracing implementation by
preventing long running requests from appearing in the performance
results.

Secondly, by recording this as an issue, there's a suggestion that
there's an action we can take to resolve it. Unfortunately, all the
times this exception occurs is when the NOMIS API is slow, and therefore
we have no control over it. I also notice that we're only seeing this
problem in staging and UAT.