Add reports:person_escort_record_quality
This adds a new Rake task which outputs a CSV of the person escort
record quality, using the `Reports::PersonEscortRecordQuality` class.

I've put it in a reports namespace assuming we will continue to add
other reporting tasks to that same namespace.

I've had to ignore a couple of `RSpec/LetSetup` Rubocop rules due to
https://github.com/rubocop/rubocop-rspec/issues/830.