Allow including journey information on moves
This should allow clients to get location information of different
journeys of a move. We need this because the timeline events sometimes
contain references to journeys which are cancelled, and therefore we
have no way of including the `to_location` and `from_location` fields of
those journeys. This means the timeline view on the frontend is
sometimes missing information (this is the case where a journey has a
different `to_location` or `from_location` of the move and therefore
isn't included).

This might have a performance impact on the endpoint, but we can
monitor closely is this becomes a problem. I don't expect many moves to
have multiple journeys.