Warn when NOMIS rejects court dates due to location mismatch
At the moment we get an exception whenever a court date is rejected
because of a location mismatch (where NOMIS doesn’t understand that a
prisoner may be in a different location in the future). Rather than
reporting this as an error in Sentry, we want to call it out as its own
type of warning so we can focus on other errors and get a better idea of
how much of a problem this one is.