Consider completed PERs as 'ready for transit'
This updates the filter for moves such that a completed per is
considered ready for transit, rather than just a confirmed one. This
change means that moves with a completed, but not confirmed, PER will
show in the "ready for transit" section rather than the "incomplete"
section.

The old behaviour makes sense when PERs used to have to be confirmed,
but now that PERs are automatically confirmed when a move starts, it
doesn't make sense to keep the move in the incomplete state until that
point.