Allow including to_location in timeline events
This follows on from
https://github.com/ministryofjustice/hmpps-book-secure-move-api/pull/1554
where we added support for including journey information when requesting
a move. Unfortunately this doesn't cover the case where an event might
include a location which isn't referenced in any journey, and therefore
will be missing from the relationships.

This change ensures the locations are included in the output and
therefore we can show the full event information on the frontend.