Use events to transition move status
This modifies the import process for moves without an ending state to
rely on the existing event which should already be there. This change is
so we can support cancelled or rejected moves that do more than just
change the status of the move, they also assign a cancellation_reason
field.