Ensure moves and journeys locations are different
This adds validation on moves and journeys to ensure that the
`from_location` and `to_location` fields are different. It shouldn't be
valid for a move to have the same source/destination location.

We may need to hold off on deploying this until both of our suppliers
are aware and comfortable of the validation.