Add Imports::DeleteEvents
This is an import task which deletes events. Sometimes the suppliers
will send up duplicate events and so we need a way of deleting them.
This is quite a crude method of simply destroying the records, but it
follows the process for the other import tasks where we're trusting the
data from Serco and not auditing it.