Add cancellation reason for database corrections
We need to manually fix the data on moves which were left uncancelled,
however if the supplier has lost information such as the cancellation
reason we need to pick one. Rather than using `other`, we can be
explicit and say that we don't know the correct cancellation reason and
that the data is not 100% accurate.