Expand scope of supplier locations endpoint
Currently, this endpoint only returns locations for a supplier if
there's a move booked from that location for the associated supplier.
This means when new locations are added and assigned to a supplier, it
won't be included in this endpoint until a move has been booked from it.

We'd like to expand the scope of this so we have this existing
behaviour, plus we return any locations where moves may (in the future)
be booked from, allowing the suppliers to be aware of all possible
locations. In particular, this helps our supplier users (for example
contract delivery managers) see all the locations in the system.