Allow suppliers to manage moves of other suppliers
This is only allows if the move is between locations owned by different
suppliers. For example, if location A is managed by supplier A and
location B is managed by supplier B, and there is a move booked which
goes between location A and B, both suppliers should have permission to
manage this move.

This functionality allows us to support "supplier to supplier" moves
where the person is handed over to a different supplier during the
course of the move.