Add date field to journeys
This will be used to store the date the journey occurred on, allowing us
to better support moves that occur over multiple days. At the moment the
field is nullable, but in the future we should expect to make this a
non-nullable field, since we can take the date from the move if it's not
specified by the suppliers.