Add PersonnelNumberValidation
This validates a model based on the values of the personnel fields for
either suppliers or police. It requires that, if the fields exist on the
model, that at least one of them is set (either for suppliers or for
police) and that either the singular or plural version of the field is
used.