Include journeys when finding moves
This modifies how the move finder works to search using the journeys
associated with a move, as well as the move itself. This is necessary to
support moves which happen across multiple days with a location in the
middle.

The difference is that we now include the date of the journeys and the
to/from location of the journeys when performing the search.