Apply journey filters together
This slightly changes the logic of how the journey filters are applied
to query the journeys (with the date and location where clause)
together. This means we avoid a move appearing for the wrong day if
there happens to be a journey for that day even if the location is
incorrect.