Prevent allocation moves date being updated
We currently don't allow the date of an allocation to be modified after
it's been created, but the date of the moves can be changed. This can
cause an mismatch between the allocation and the moves which leads to
the allocation not appearing in the right place in the frontend.