Fix finding moves with journeys
This follows up
https://github.com/ministryofjustice/hmpps-book-secure-move-api/pull/1791
to clarify some logic which was missing from the original
implementation. Specifically, this prevents moves from appearing when
they shouldn't, as rather than combining the move date/locations and the
journey date/locations, we ignore the move details if journeys are there
(this does require trusting the journey details are correct for
multi-day moves).

I've added tests for the 12 different scenarious and when the moves
should or should not appear.