Refresh allocation status when cancelling a move
Cancelling a move can affect the status of an allocation as only moves
which are not cancelled are used to determine the status. Without this,
you can end up in a state where an allocation is left unfilled if the
one of the moves ends up being cancelled before the other moves start.