Add index on journey state and date
This should hopefully improve the performance of the new moves finder
where it looks at journeys as well as moves. The query plan suggests it
has to be a scan of the table on these fields, so putting in an index
should help that.

```
    ->  Bitmap Heap Scan on journeys journeys_3  (cost=57.85..2531.25 rows=1 width=16) (actual time=0.031..0.032 rows=0 loops=1)
          Recheck Cond: (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid)
"          Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (date >= '2022-01-01'::date) AND (date <= '2022-01-01'::date))"
          ->  Bitmap Index Scan on index_journeys_on_from_location_id  (cost=0.00..57.85 rows=723 width=0) (actual time=0.028..0.028 rows=0 loops=1)
                Index Cond: (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid)
```