Update assessment status based on required questions
Currently we check if all of the questions have been responded to update
the status of the assessment to complete. This isn't quite right as some
questions are not required, and shouldn't prevent the assessment from
being marked as complete.

Specifically, this is required to support the date of arrest optional
question we're adding to the frameworks:
https://github.com/ministryofjustice/hmpps-book-secure-move-frameworks/pull/63